+++ src/graph_rag/ingest/scip/__init__.py
@@ -1,0 +1,13 @@
+from .scip_document import ScipDocument
+from .scip_occurrence import ScipOccurrence
+from .scip_reader import ScipReader
+from .scip_symbol import ScipSymbol
+from .scip_symbol_parser import ScipSymbolParser
+
+__all__ = [
+    "ScipDocument",
+    "ScipOccurrence",
+    "ScipReader",
+    "ScipSymbol",
+    "ScipSymbolParser",
+]